i'm not sure if this is working

+++ main.py
@@ -4,32 +4,41 @@ import time, json
 class Wallet():
 
     def __init__(self):
-        self.cash = 10000
+        self.cash = 10000 #digital fake cash for simulation purposes
+
+    def checkIfIn(self, ID):
+        with open('data.json', 'r+') as file:
+            data = json.load(file) # load in json file
+
+            try:
+                str = data["shares"][ID]
+                return True #if ID exists in json file, return True
+            except KeyError:
+                return False# if ID does not exist in json file, return False
 
     def buy(self, ID):
         obj = ShareObj(ID)
-        obj.refresh()
+        obj.refresh() # update prices
 
-        self.cash -= float(obj.getPrice())
+        if(self.checkIfIn(ID) == False): # if share is not already in portfolio
+            self.cash -= float(obj.getPrice()) #subtract from digital wallet
+            with open('data.json', 'r+') as file: #and update json to reflect stock purchase
+                # read file
+                data = json.load(file)
+                # wipe file
+                file.seek(0)
+                file.truncate()
 
-        with open('data.json', 'r+') as file:
+                # add new price to appropriate id
+                data["shares"].append({
+                    "id": obj.getID(),
+                    "price": obj.getPrice()
+                })
 
-            # read file
-            data = json.load(file)
+                data = str(data).replace("'", '"')
 
-            # wipe file
-            file.seek(0)
-            file.truncate()
-
-            # add new price to appropriate id
-            data['shares'][ID].append({
-                "price": obj.getPrice()
-            })
-
-            data = str(data).replace("'", '"')
-
-            file.write(data)
-            file.close()
+                file.write(data) #write
+                file.close()
 
     def sell(self, ID):
         obj = ShareObj(ID)
@@ -61,11 +70,15 @@ class ShareObj(object):
     def getID(self):
         return self.id
 
+    def setID(self, newID):
+        self.id = newID
+        self.refresh()
+
     def refresh(self):
         self.share.refresh()
 
 w = Wallet()
-shre = ShareObj("AAPL")
+shr = ShareObj("TSLA")
 
 percentChange = 0.15
 
@@ -73,20 +86,20 @@ print("Initializing...\n")
 
 while(True):
     print("Wallet:\t\t\t",w.getCash())
-    print("Percent Change:\t\t", shre.getChange(),"\n")
+    print("Percent Change:\t\t", shr.getChange(),"\n")
 
-    if(float(shre.getChange()) >= percentChange):
+    if(float(shr.getChange()) >= percentChange):
         print("Buy")
         print("Wallet before buy:\t",w.getCash())
-        w.buy(shre.getID())
+        w.buy(shr.getID())
         print("Wallet after buy:\t",w.getCash())
-    elif(float(shre.getChange()) <= (-1*percentChange)):
+    elif(float(shr.getChange()) <= (-1*percentChange)):
         print("Sell")
         print("Waller before sell:\t",w.getCash())
-        w.sell(shre.getID())
+        w.sell(shr.getID())
         print("Wallet before sell:t\t",w.getCash())
     else:
         print("Do Nothing")
     print("\n")
-    shre.refresh
+    shr.refresh
     time.sleep(10)
